bindings: complete Phase B/C rewiring, codec extensions, and doc cleanup
Phase A — C binding layer separation:
- bindings/c/: new aggregate convenience layer (zlink_c.h + zlink_c.c)
- core/include/zlink.h: aggregate block removed, *_part substrate only
- libzlink.so.5.3.0 + libzlink_c.so deployed to all 8 binding native dirs

Phase B — all 7 language bindings rewired to *_part substrate:
- cpp/dotnet/go/java/node/python/rust: aggregate→*_part, Gate 1-3 all pass
- Non-canonical callback samples removed (dealer_router/pair/pubsub/
  request_reply/spot) from cpp/go/node/python/rust
- Perf runners updated (warmup removal, transport policy, metric correctness)

Phase C — codec extensions (protobuf/json/messagepack) for all 6 bindings:
- bindings/{cpp,dotnet,go,java,node,python,rust}/codecs/: scaffold + tests
- Java: zlink-codec-{protobuf,json,messagepack} + zlink-ext-netty (all pass)
- Java Message.java: ByteBuf removed; wrapDirect(ByteBuffer) public;
  NettyMessageAdapter rewritten without reflection (4/4 new contract tests)

doc/guide cleanup:
- core/samples/ → bindings/c/samples/ (26 files)
- stream_callback_sample → stream_packet_callback_sample
- dealer_router_callback_sample / spot_callback_sample sections removed

+++ bindings/rust/src/flags.rs
@@ -5,7 +5,7 @@ pub struct SendFlags(u32);
 
 impl SendFlags {
     pub const NONE: Self = Self(0);
-    pub const DONT_WAIT: Self = Self(ffi::ZLINK_DONTWAIT as u32);
+    pub const DONT_WAIT: Self = Self(ffi::ZLINK_DONTWAIT);
 
     pub const fn bits(self) -> u32 {
         self.0
@@ -23,7 +23,7 @@ pub struct RecvFlags(u32);
 
 impl RecvFlags {
     pub const NONE: Self = Self(0);
-    pub const DONT_WAIT: Self = Self(ffi::ZLINK_DONTWAIT as u32);
+    pub const DONT_WAIT: Self = Self(ffi::ZLINK_DONTWAIT);
 
     pub const fn bits(self) -> u32 {
         self.0
